requirements.txt
pandas
pandas-ta
requests
python-telegram-bot